fix: 식재료(Ingredient) 수량 타입을 Double로 통일하여 소수점 유실 버그 수정
- Ingredient 엔티티 및 관련 DTO의 quantity 타입을 Integer에서 Double로 변경
- IngredientRequestDto의 수량 유효성 검사를 @Positive로 수정하여 소수점 허용
- ShoppingItemService의 냉장고 이관 로직(moveCheckedItemsToFridge)에서 intValue() 강제 형변환 제거
- 장바구니 항목(Double)을 냉장고로 옮길 때 수량이 정확히 유지되도록 개선

+++ src/main/java/com/example/Naengbuhae/domain/Ingredient.java
@@ -23,7 +23,8 @@ public class Ingredient {
     @Column(nullable = false)
     private String name; // 식재료 이름 (예: 계란)
 
-    private Integer quantity; // 수량 (예: 10)
+    // 기존: private Integer quantity;
+    private Double quantity; // 수량 (소수점 가능, 예: 1.5)
 
     private LocalDate expirationDate; // 유통기한 (예: 2026-04-15)
 
@@ -39,7 +40,7 @@ public class Ingredient {
     @Column(nullable = false)
     private LocalDate purchaseDate; // 구매일 (예: 2026-04-15)
 
-    public Ingredient(User user, String name, Integer quantity, LocalDate expirationDate,
+    public Ingredient(User user, String name, Double quantity, LocalDate expirationDate,
                       String category, String unit, String storage, LocalDate purchaseDate) {
         this.user = user;
         this.name = name;
